pyproject.toml
[project]
name = "eimer-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Karl Blumenthal", email = "dev@profanta.de" }
]
requires-python = ">=3.12"
dependencies = [
    "jpysocket>=1.1.5",
    "pillow>=11.1.0",
    "rich-structlog>=0.4.1",
    "rich>=14.0.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
]

[project.scripts]
eimer-client = "eimer_client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
